Create a mini router in React- use history api
Used history api to create a mini router in React. Works
fine for 1 pass of front navigation and backward navigation. But
an Insidious bug entered : The next time you press forward,
Back button disappears, and only appears when i click on the
screen of the browser (even when I didn't write any code with
click event listener for history api). Wasted a day for it, but
learned about HISTORY API and browser LOCATION property which are the
main js APIs used in routing of Single Page Web Applications.