Fix revoke-compounding of Refresh Tokens
When I revoke all Refresh tokens of a user when someone tries to hack
using an Invalid Refresh token, then when the user tries to login
from another browser, it sends the same revoked Refresh token stored
in the cookie (if cookie not expired), which then triggers Revoking of all
RTs again. The cycle continues. To prevent this, revoke only the hacking
user's Refresh token.