Create "Beyond Border" feature for canvas resizing
When making the canvas bigger, the drawings need to hide and show
according to how much visual css pixels are there in the viewable
canvas. This is achieved by using a big Offscreen canvas where all the
drawings are permanently stored, and the undo-redo features are done on
this canvas - to improve performance too. This full canvas is drawn on
the viewable canvas in each frame when slider moves. Thus causing Beyond
border feature.